server/requirements.txt
click==8.0.1
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
Werkzeug==2.0.1
